requirements.txt
urllib2
requests
beautifulsoup4
phonenumbers
tweepy
python-decouple
facebook-sdk
mechanize
